requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.6
python-dotenv>=0.21.0
typing-extensions>=4.0.0

# Vector database
qdrant-client>=1.7.0

# Local LLM
vllm>=0.8.5

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0